refactor: extract core extension projected contract finalizer
